backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.10,<3.0
python-multipart>=0.0.9
httpx>=0.27,<1.0
reportlab>=4.2,<5.0
sqlalchemy>=2.0.35,<3.0
psycopg[binary]>=3.2,<4.0
pypdf>=5.0,<6.0
python-docx>=1.1,<2.0
azure-ai-inference>=1.0.0b9,<2.0
azure-core>=1.30,<2.0
aiohttp>=3.10,<4.0
alembic>=1.13,<2.0
python-dotenv>=1.0,<2.0
